Data.SBV.Internals

Description

Low level functions to access the SBV infrastructure, for developers who want to build further tools on top of SBV. End-users of the library should not need to use this module.

NB. There are various coding invariants in SBV that are maintained throughout the code. Indiscriminate use of functions in this module can break those invariants. So, you are on your own if you do utilize the functions here. (Unfortunately, what exactly those invariants are is a very good but also a very difficult question to answer!)

Running symbolic programs manually

data Result

Result of running a symbolic computation

Constructors

Result

Fields

reskinds :: Set Kind

kinds used in the program

resTraces :: [(String, CV)]

quick-check counter-example information (if any)

resObservables :: [(String, CV -> Bool, SV)]

observable expressions (part of the model)

resUISegs :: [(String, [String])]

uninterpeted code segments

resInputs :: ([(Quantifier, NamedSymVar)], [NamedSymVar])

inputs (possibly existential) + tracker vars

resConsts :: [(SV, CV)]

constants

resTables :: [((Int, Kind, Kind), [SV])]

tables (automatically constructed) (tableno, index-type, result-type) elts

resArrays :: [(Int, ArrayInfo)]

arrays (user specified)

resUIConsts :: [(String, SBVType)]

uninterpreted constants

resAxioms :: [(String, [String])]

axioms

resAsgns :: SBVPgm

assignments

resConstraints :: Seq (Bool, [(String, String)], SV)

additional constraints (boolean)

resAssertions :: [(String, Maybe CallStack, SV)]

assertions

resOutputs :: [SV]

outputs

data SBVRunMode

Different means of running a symbolic piece of code

Constructors

SMTMode QueryContext IStage Bool SMTConfig
In regular mode, with a stage. Bool is True if this is SAT.

CodeGen
Code generation mode.

Concrete (Maybe (Bool, [((Quantifier, NamedSymVar), Maybe CV)]))
Concrete simulation mode, with given environment if any. If Nothing: Random.

data IStage

Stage of an interactive run

Constructors

ISetup

ISafe

IRun

data QueryContext

Query execution context

Constructors

QueryInternal
Triggered from inside SBV

QueryExternal
Triggered from user code

Solver capabilities

data SolverCapabilities

Translation tricks needed for specific capabilities afforded by each solver

Constructors

SolverCapabilities

Fields

supportsQuantifiers :: Bool

Supports SMT-Lib2 style quantifiers?

supportsUninterpretedSorts :: Bool

Supports SMT-Lib2 style uninterpreted-sorts

supportsUnboundedInts :: Bool

Supports unbounded integers?

supportsReals :: Bool

Supports reals?

supportsApproxReals :: Bool

Supports printing of approximations of reals?

supportsIEEE754 :: Bool

Supports floating point numbers?

supportsSets :: Bool

Supports set operations?

supportsOptimization :: Bool

Supports optimization routines?

supportsPseudoBooleans :: Bool

Supports pseudo-boolean operations?

supportsCustomQueries :: Bool

Supports interactive queries per SMT-Lib?

supportsGlobalDecls :: Bool

Supports global declarations? (Needed for push-pop.)

supportsDataTypes :: Bool

Supports datatypes?

supportsFlattenedModels :: Maybe [String]

Supports flattened model output? (With given config lines.)

Internal structures useful for low-level programming

type SBool = SBV Bool

A symbolic boolean/bit

type SWord8 = SBV Word8

8-bit unsigned symbolic value

type SWord16 = SBV Word16

16-bit unsigned symbolic value

type SWord32 = SBV Word32

32-bit unsigned symbolic value

type SWord64 = SBV Word64

64-bit unsigned symbolic value

type SInt8 = SBV Int8

8-bit signed symbolic value, 2's complement representation

type SInt16 = SBV Int16

16-bit signed symbolic value, 2's complement representation

type SInt32 = SBV Int32

32-bit signed symbolic value, 2's complement representation

type SInt64 = SBV Int64

64-bit signed symbolic value, 2's complement representation

type SInteger = SBV Integer

Infinite precision signed symbolic value

type SReal = SBV AlgReal

Infinite precision symbolic algebraic real value

type SFloat = SBV Float

IEEE-754 single-precision floating point numbers

type SDouble = SBV Double

IEEE-754 double-precision floating point numbers

type SChar = SBV Char

A symbolic character. Note that, as far as SBV's symbolic strings are concerned, a character is currently an 8-bit unsigned value, corresponding to the ISO-8859-1 (Latin-1) character set: http://en.wikipedia.org/wiki/ISO/IEC_8859-1. A Haskell Char, on the other hand, is based on unicode. Therefore, there isn't a 1-1 correspondence between a Haskell character and an SBV character for the time being. This limitation is due to the SMT-solvers only supporting this particular subset. However, there is a pending proposal to add support for unicode, and SBV will track these changes to have full unicode support as solvers become available. For details, see: http://smtlib.cs.uiowa.edu/theories-UnicodeStrings.shtml

type SString = SBV String

A symbolic string. Note that a symbolic string is not a list of symbolic characters, that is, it is not the case that SString = [SChar], unlike what one might expect following Haskell strings. An SString is a symbolic value of its own, of possibly arbitrary but finite length, and internally processed as one unit as opposed to a fixed-length list of characters.

type SList a = SBV [a]

A symbolic list of items. Note that a symbolic list is not a list of symbolic items, that is, it is not the case that SList a = [a], unlike what one might expect following haskell lists/sequences. An SList is a symbolic value of its own, of possibly arbitrary but finite length, and internally processed as one unit as opposed to a fixed-length list of items. Note that lists can be nested, i.e., we do allow lists of lists of ... items.

type SEither a b = SBV (Either a b)

Symbolic Either

type SMaybe a = SBV (Maybe a)

Symbolic Maybe

type STuple a b = SBV (a, b)

Symbolic 2-tuple. NB. STuple and STuple2 are equivalent.

type STuple2 a b = SBV (a, b)

Symbolic 2-tuple. NB. STuple and STuple2 are equivalent.

type STuple3 a b c = SBV (a, b, c)

Symbolic 3-tuple.

type STuple4 a b c d = SBV (a, b, c, d)

Symbolic 4-tuple.

type STuple5 a b c d e = SBV (a, b, c, d, e)

Symbolic 5-tuple.

type STuple6 a b c d e f = SBV (a, b, c, d, e, f)

Symbolic 6-tuple.

type STuple7 a b c d e f g = SBV (a, b, c, d, e, f, g)

Symbolic 7-tuple.

type STuple8 a b c d e f g h = SBV (a, b, c, d, e, f, g, h)

Symbolic 8-tuple.

data RCSet a

A RCSet is either a regular set or a set given by its complement from the corresponding universal set.

Constructors

RegularSet (Set a)

ComplementSet (Set a)

type SSet a = SBV (RCSet a)

Symbolic Set. Note that we use RCSet, which supports both regular sets and complements, i.e., those obtained from the universal set (of the right type) by removing elements.

nan :: Floating a => a

Not-A-Number for Double and Float. Surprisingly, Haskell Prelude doesn't have this value defined, so we provide it here.

infinity :: Floating a => a

Infinity for Double and Float. Surprisingly, Haskell Prelude doesn't have this value defined, so we provide it here.

sNaN :: (Floating a, SymVal a) => SBV a

Symbolic variant of Not-A-Number. This value will inhabit both SDouble and SFloat.

sInfinity :: (Floating a, SymVal a) => SBV a

Symbolic variant of infinity. This value will inhabit both SDouble and SFloat.

data RoundingMode

Rounding mode to be used for the IEEE floating-point operations. Note that Haskell's default is RoundNearestTiesToEven. If you use a different rounding mode, then the counter-examples you get may not match what you observe in Haskell.

Constructors

RoundNearestTiesToEven
Round to nearest representable floating point value. If precisely at half-way, pick the even number. (In this context, even means the lowest-order bit is zero.)

RoundNearestTiesToAway
Round to nearest representable floating point value. If precisely at half-way, pick the number further away from 0. (That is, for positive values, pick the greater; for negative values, pick the smaller.)

RoundTowardPositive
Round towards positive infinity. (Also known as rounding-up or ceiling.)

RoundTowardNegative
Round towards negative infinity. (Also known as rounding-down or floor.)

RoundTowardZero
Round towards zero. (Also known as truncation.)

type SRoundingMode = SBV RoundingMode

The symbolic variant of RoundingMode

sRoundNearestTiesToEven :: SRoundingMode

Symbolic variant of RoundNearestTiesToEven

sRoundNearestTiesToAway :: SRoundingMode

Symbolic variant of RoundNearestTiesToAway

sRoundTowardPositive :: SRoundingMode

Symbolic variant of RoundTowardPositive

sRoundTowardNegative :: SRoundingMode

Symbolic variant of RoundTowardNegative

sRoundTowardZero :: SRoundingMode

Symbolic variant of RoundTowardZero

sRNE :: SRoundingMode

Alias for sRoundNearestTiesToEven

sRNA :: SRoundingMode

Alias for sRoundNearestTiesToAway

sRTP :: SRoundingMode

Alias for sRoundTowardPositive

sRTN :: SRoundingMode

Alias for sRoundTowardNegative

sRTZ :: SRoundingMode

Alias for sRoundTowardZero

class (HasKind a, Typeable a) => SymVal a where

A SymVal is a potential symbolic value that can be created instances of to be fed to a symbolic program.

Minimal complete definition

Nothing

Methods

mkSymVal :: MonadSymbolic m => Maybe Quantifier -> Maybe String -> m (SBV a)

Generalization of mkSymVal

default mkSymVal :: (MonadSymbolic m, Read a, Data a) => Maybe Quantifier -> Maybe String -> m (SBV a)

literal :: a -> SBV a

Turn a literal constant to symbolic

default literal :: Show a => a -> SBV a

fromCV :: CV -> a

Extract a literal, from a CV representation

default fromCV :: Read a => CV -> a

isConcretely :: SBV a -> (a -> Bool) -> Bool

Does it concretely satisfy the given predicate?

forall :: MonadSymbolic m => String -> m (SBV a)

Generalization of forall

forall_ :: MonadSymbolic m => m (SBV a)

Generalization of forall_

mkForallVars :: MonadSymbolic m => Int -> m [SBV a]

Generalization of mkForallVars

exists :: MonadSymbolic m => String -> m (SBV a)

Generalization of exists

exists_ :: MonadSymbolic m => m (SBV a)

Generalization of exists_

mkExistVars :: MonadSymbolic m => Int -> m [SBV a]

Generalization of mkExistVars

free :: MonadSymbolic m => String -> m (SBV a)

Generalization of free

free_ :: MonadSymbolic m => m (SBV a)

Generalization of free_

mkFreeVars :: MonadSymbolic m => Int -> m [SBV a]

Generalization of mkFreeVars

symbolic :: MonadSymbolic m => String -> m (SBV a)

Generalization of symbolic

symbolics :: MonadSymbolic m => [String] -> m [SBV a]

Generalization of symbolics

unliteral :: SBV a -> Maybe a

Extract a literal, if the value is concrete

isConcrete :: SBV a -> Bool

Is the symbolic word concrete?

isSymbolic :: SBV a -> Bool

Is the symbolic word really symbolic?

data CV

CV represents a concrete word of a fixed size: For signed words, the most significant digit is considered to be the sign.

Constructors

CV

Fields

_cvKind :: !Kind

cvVal :: !CVal

data CVal

A constant value

Constructors

CAlgReal !AlgReal
Algebraic real

CInteger !Integer
Bit-vector/unbounded integer

CFloat !Float
Float

CDouble !Double
Double

CChar !Char
Character

CString !String
String

CList ![CVal]
List

CSet !(RCSet CVal)
Set. Can be regular or complemented.

CUserSort !(Maybe Int, String)
Value of an uninterpreted/user kind. The Maybe Int shows index position for enumerations

CTuple ![CVal]
Tuple

CMaybe !(Maybe CVal)
Maybe

CEither !(Either CVal CVal)
Disjoint union

data AlgReal

Algebraic reals. Note that the representation is left abstract. We represent rational results explicitly, while the roots-of-polynomials are represented implicitly by their defining equation

Constructors

AlgRational Bool Rational
bool says it's exact (i.e., SMT-solver did not return it with ? at the end.)

AlgPolyRoot (Integer, AlgRealPoly) (Maybe String)
which root of this polynomial and an approximate decimal representation with given precision, if available

newtype AlgRealPoly

A univariate polynomial, represented simply as a coefficient list. For instance, "5x^3 + 2x - 5" is represented as [(5, 3), (2, 1), (-5, 0)]

Constructors

AlgRealPoly [(Integer, Integer)]

data ExtCV

A simple expression type over extendent values, covering infinity, epsilon and intervals.

Constructors

Infinite Kind

Epsilon Kind

Interval ExtCV ExtCV

BoundedCV CV

AddExtCV ExtCV ExtCV

MulExtCV ExtCV ExtCV

data GeneralizedCV

A generalized CV allows for expressions involving infinite and epsilon values/intervals Used in optimization problems.

Constructors

ExtendedCV ExtCV

RegularCV CV

isRegularCV :: GeneralizedCV -> Bool

Is this a regular CV?

cvSameType :: CV -> CV -> Bool

Are two CV's of the same type?

cvToBool :: CV -> Bool

Convert a CV to a Haskell boolean (NB. Assumes input is well-kinded)

mkConstCV :: Integral a => Kind -> a -> CV

Create a constant word from an integral.

liftCV2 :: (AlgReal -> AlgReal -> b) -> (Integer -> Integer -> b) -> (Float -> Float -> b) -> (Double -> Double -> b) -> (Char -> Char -> b) -> (String -> String -> b) -> ([CVal] -> [CVal] -> b) -> ([CVal] -> [CVal] -> b) -> (Maybe CVal -> Maybe CVal -> b) -> (Either CVal CVal -> Either CVal CVal -> b) -> ((Maybe Int, String) -> (Maybe Int, String) -> b) -> CV -> CV -> b

Lift a binary function through a CV.

mapCV :: (AlgReal -> AlgReal) -> (Integer -> Integer) -> (Float -> Float) -> (Double -> Double) -> (Char -> Char) -> (String -> String) -> ((Maybe Int, String) -> (Maybe Int, String)) -> CV -> CV

Map a unary function through a CV.

mapCV2 :: (AlgReal -> AlgReal -> AlgReal) -> (Integer -> Integer -> Integer) -> (Float -> Float -> Float) -> (Double -> Double -> Double) -> (Char -> Char -> Char) -> (String -> String -> String) -> ((Maybe Int, String) -> (Maybe Int, String) -> (Maybe Int, String)) -> CV -> CV -> CV

Map a binary function through a CV.

data SV

A symbolic word, tracking it's signedness and size.

Constructors

SV !Kind !NodeId

trueSV :: SV

Constant True as an SV. Note that this value always occupies slot -1.

falseSV :: SV

Constant False as an SV. Note that this value always occupies slot -2.

trueCV :: CV

Constant True as a CV. We represent it using the integer value 1.

falseCV :: CV

Constant False as a CV. We represent it using the integer value 0.

normCV :: CV -> CV

Normalize a CV. Essentially performs modular arithmetic to make sure the value can fit in the given bit-size. Note that this is rather tricky for negative values, due to asymmetry. (i.e., an 8-bit negative number represents values in the range -128 to 127; thus we have to be careful on the negative side.)

data SVal

The Symbolic value. Either a constant (Left) or a symbolic value (Right Cached). Note that caching is essential for making sure sharing is preserved.

Constructors

SVal !Kind !(Either CV (Cached SV))

sTrue :: SBool

Symbolic True

sFalse :: SBool

Symbolic False

sNot :: SBool -> SBool

Symbolic boolean negation

(.&&) :: SBool -> SBool -> SBool infixr 3

Symbolic conjunction

(.||) :: SBool -> SBool -> SBool infixr 2

Symbolic disjunction

(.<+>) :: SBool -> SBool -> SBool infixl 6

Symbolic logical xor

(.~&) :: SBool -> SBool -> SBool infixr 3

Symbolic nand

(.~|) :: SBool -> SBool -> SBool infixr 2

Symbolic nor

(.=>) :: SBool -> SBool -> SBool infixr 1

Symbolic implication

(.<=>) :: SBool -> SBool -> SBool infixr 1

Symbolic boolean equivalence

sAnd :: [SBool] -> SBool

Generalization of and

sOr :: [SBool] -> SBool

Generalization of or

sAny :: (a -> SBool) -> [a] -> SBool

Generalization of any

sAll :: (a -> SBool) -> [a] -> SBool

Generalization of all

fromBool :: Bool -> SBool

Conversion from Bool to SBool

newtype SBV a

The Symbolic value. The parameter a is phantom, but is extremely important in keeping the user interface strongly typed.

Constructors

SBV

Fields

unSBV :: SVal

newtype NodeId

A symbolic node id

Constructors

NodeId Int

mkSymSBV :: forall a m. MonadSymbolic m => Maybe Quantifier -> Kind -> Maybe String -> m (SBV a)

Generalization of mkSymSBV

data ArrayContext

The context of a symbolic array as created

Constructors

ArrayFree (Maybe SV)
A new array, the contents are initialized with the given value, if any

ArrayMutate ArrayIndex SV SV
An array created by mutating another array at a given cell

ArrayMerge SV ArrayIndex ArrayIndex
An array created by symbolically merging two other arrays

type ArrayInfo = (String, (Kind, Kind), ArrayContext)

Representation for symbolic arrays

class SymArray array where

Flat arrays of symbolic values An array a b is an array indexed by the type SBV a, with elements of type SBV b.

If a default value is supplied, then all the array elements will be initialized to this value. Otherwise, they will be left unspecified, i.e., a read from an unwritten location will produce an uninterpreted constant.

While it's certainly possible for user to create instances of SymArray, the SArray and SFunArray instances already provided should cover most use cases in practice. Note that there are a few differences between these two models in terms of use models:

SArray produces SMTLib arrays, and requires a solver that understands the array theory. SFunArray is internally handled, and thus can be used with any solver. (Note that all solvers except abc support arrays, so this isn't a big decision factor.)

For both arrays, if a default value is supplied, then reading from uninitialized cell will return that value. If the default is not given, then reading from uninitialized cells is still OK for both arrays, and will produce an uninterpreted constant in both cases.

Only SArray supports checking equality of arrays. (That is, checking if an entire array is equivalent to another.) SFunArrays cannot be checked for equality. In general, checking wholesale equality of arrays is a difficult decision problem and should be avoided if possible.

Only SFunArray supports compilation to C. Programs using SArray will not be accepted by the C-code generator.

You cannot use quickcheck on programs that contain these arrays. (Neither SArray nor SFunArray.)

With SArray, SBV transfers all array-processing to the SMT-solver. So, it can generate programs more quickly, but they might end up being too hard for the solver to handle. With SFunArray, SBV only generates code for individual elements and the array itself never shows up in the resulting SMTLib program. This puts more onus on the SBV side and might have some performance impacts, but it might generate problems that are easier for the SMT solvers to handle.

As a rule of thumb, try SArray first. These should generate compact code. However, if the backend solver has hard time solving the generated problems, switch to SFunArray. If you still have issues, please report so we can see what the problem might be!

Minimal complete definition

readArray, writeArray, mergeArrays, (newArray_, newArray | newArrayInState)

Methods

newArray_ :: (MonadSymbolic m, HasKind a, HasKind b) => Maybe (SBV b) -> m (array a b)

Generalization of newArray_

newArray :: (MonadSymbolic m, HasKind a, HasKind b) => String -> Maybe (SBV b) -> m (array a b)

Generalization of newArray

readArray :: array a b -> SBV a -> SBV b

Read the array element at a

writeArray :: SymVal b => array a b -> SBV a -> SBV b -> array a b

Update the element at a to be b

mergeArrays :: SymVal b => SBV Bool -> array a b -> array a b -> array a b

Merge two given arrays on the symbolic condition Intuitively: mergeArrays cond a b = if cond then a else b. Merging pushes the if-then-else choice down on to elements

newArrayInState :: (HasKind a, HasKind b) => Maybe String -> Maybe (SBV b) -> State -> IO (array a b)

Internal function, not exported to the user

newtype SFunArray a b

Arrays implemented internally, without translating to SMT-Lib functions:

Internally handled by the library and not mapped to SMT-Lib, hence can be used with solvers that don't support arrays. (Such as abc.)

Reading from an unintialized value is OK. If the default value is given in newArray, it will be the result. Otherwise, the read yields an uninterpreted constant.

Cannot check for equality of arrays.

Can be used in code-generation (i.e., compilation to C).

Can not quick-check theorems using SFunArray values

Typically faster as it gets compiled away during translation.

Constructors

SFunArray

Fields

unSFunArray :: SFunArr

newtype SArray a b

Arrays implemented in terms of SMT-arrays: http://smtlib.cs.uiowa.edu/theories-ArraysEx.shtml

Maps directly to SMT-lib arrays

Reading from an unintialized value is OK. If the default value is given in newArray, it will be the result. Otherwise, the read yields an uninterpreted constant.

Can check for equality of these arrays

Cannot be used in code-generation (i.e., compilation to C)

Cannot quick-check theorems using SArray values

Typically slower as it heavily relies on SMT-solving for the array theory

Constructors

SArray

Fields

unSArray :: SArr

sbvToSV :: State -> SBV a -> IO SV

Convert a symbolic value to a symbolic-word

sbvToSymSV :: MonadSymbolic m => SBV a -> m SV

Generalization of sbvToSymSW

forceSVArg :: SV -> IO ()

Forcing an argument; this is a necessary evil to make sure all the arguments to an uninterpreted function are evaluated before called; the semantics of uinterpreted functions is necessarily strict; deviating from Haskell's

data SBVExpr

A symbolic expression

Constructors

SBVApp !Op ![SV]

newExpr :: State -> Kind -> SBVExpr -> IO SV

Create a new expression; hash-cons as necessary

cache :: (State -> IO a) -> Cached a

Cache a state-based computation

data Cached a

We implement a peculiar caching mechanism, applicable to the use case in implementation of SBV's. Whenever we do a state based computation, we do not want to keep on evaluating it in the then-current state. That will produce essentially a semantically equivalent value. Thus, we want to run it only once, and reuse that result, capturing the sharing at the Haskell level. This is similar to the "type-safe observable sharing" work, but also takes into the account of how symbolic simulation executes.

See Andy Gill's type-safe obervable sharing trick for the inspiration behind this technique: http://ku-fpg.github.io/files/Gill-09-TypeSafeReification.pdf

Note that this is *not* a general memo utility!

uncache :: Cached SV -> State -> IO SV

Uncache a previously cached computation

uncacheAI :: Cached ArrayIndex -> State -> IO ArrayIndex

Uncache, retrieving SMT array indexes

class HasKind a where

A class for capturing values that have a sign and a size (finite or infinite) minimal complete definition: kindOf, unless you can take advantage of the default signature: This class can be automatically derived for data-types that have a Data instance; this is useful for creating uninterpreted sorts. So, in reality, end users should almost never need to define any methods.

Minimal complete definition

Nothing

Methods

kindOf :: a -> Kind

default kindOf :: (Read a, Data a) => a -> Kind

hasSign :: a -> Bool

intSizeOf :: a -> Int

isBoolean :: a -> Bool

isBounded :: a -> Bool

isReal :: a -> Bool

isFloat :: a -> Bool

isDouble :: a -> Bool

isUnbounded :: a -> Bool

isUninterpreted :: a -> Bool

isChar :: a -> Bool

isString :: a -> Bool

isList :: a -> Bool

isSet :: a -> Bool

isTuple :: a -> Bool

isMaybe :: a -> Bool

isEither :: a -> Bool

showType :: a -> String

data Op

Symbolic operations

Constructors

Plus

Times

Minus

UNeg

Abs

Quot

Rem

Equal

NotEqual

LessThan

GreaterThan

LessEq

GreaterEq

Ite

And

Or

XOr

Not

Shl

Shr

Rol Int

Ror Int

Extract Int Int

Join

LkUp (Int, Kind, Kind, Int) !SV !SV

ArrEq ArrayIndex ArrayIndex

ArrRead ArrayIndex

KindCast Kind Kind

Uninterpreted String

Label String

IEEEFP FPOp

PseudoBoolean PBOp

OverflowOp OvOp

StrOp StrOp

SeqOp SeqOp

SetOp SetOp

TupleConstructor Int

TupleAccess Int Int

EitherConstructor Kind Kind Bool

EitherIs Kind Kind Bool

EitherAccess Bool

MaybeConstructor Kind Bool

MaybeIs Kind Bool

MaybeAccess

data PBOp

Pseudo-boolean operations

Constructors

PB_AtMost Int
At most k

PB_AtLeast Int
At least k

PB_Exactly Int
Exactly k

PB_Le [Int] Int
At most k, with coefficients given. Generalizes PB_AtMost

PB_Ge [Int] Int
At least k, with coefficients given. Generalizes PB_AtLeast

PB_Eq [Int] Int
Exactly k, with coefficients given. Generalized PB_Exactly

data FPOp

Floating point operations

Constructors

FP_Cast Kind Kind SV

FP_Reinterpret Kind Kind

FP_Abs

FP_Neg

FP_Add

FP_Sub

FP_Mul

FP_Div

FP_FMA

FP_Sqrt

FP_Rem

FP_RoundToIntegral

FP_Min

FP_Max

FP_ObjEqual

FP_IsNormal

FP_IsSubnormal

FP_IsZero

FP_IsInfinite

FP_IsNaN

FP_IsNegative

FP_IsPositive

data StrOp

String operations. Note that we do not define StrAt as it translates to StrSubstr trivially.

Constructors

StrConcat
Concatenation of one or more strings

StrLen
String length

StrUnit
Unit string

StrNth
Nth element

StrSubstr
Retrieves substring of s at offset

StrIndexOf
Retrieves first position of sub in s, -1 if there are no occurrences

StrContains
Does s contain the substring sub?

StrPrefixOf
Is pre a prefix of s?

StrSuffixOf
Is suf a suffix of s?

StrReplace
Replace the first occurrence of src by dst in s

StrStrToNat
Retrieve integer encoded by string s (ground rewriting only)

StrNatToStr
Retrieve string encoded by integer i (ground rewriting only)

StrInRe RegExp
Check if string is in the regular expression

data SeqOp

Sequence operations.

Constructors

SeqConcat
See StrConcat

SeqLen
See StrLen

SeqUnit
See StrUnit

SeqNth
See StrNth

SeqSubseq
See StrSubseq

SeqIndexOf
See StrIndexOf

SeqContains
See StrContains

SeqPrefixOf
See StrPrefixOf

SeqSuffixOf
See StrSuffixOf

SeqReplace
See StrReplace

data RegExp

Regular expressions. Note that regular expressions themselves are concrete, but the match function from the RegExpMatchable class can check membership against a symbolic string/character. Also, we are preferring a datatype approach here, as opposed to coming up with some string-representation; there are way too many alternatives already so inventing one isn't a priority. Please get in touch if you would like a parser for this type as it might be easier to use.

Constructors

Literal String
Precisely match the given string

All
Accept every string

None
Accept no strings

Range Char Char
Accept range of characters

Conc [RegExp]
Concatenation

KStar RegExp
Kleene Star: Zero or more

KPlus RegExp
Kleene Plus: One or more

Opt RegExp
Zero or one

Loop Int Int RegExp
From n repetitions to m repetitions

Union [RegExp]
Union of regular expressions

Inter RegExp RegExp
Intersection of regular expressions

type NamedSymVar = (SV, String)

NamedSymVar pairs symbolic values and user given/automatically generated names

getTableIndex :: State -> Kind -> Kind -> [SV] -> IO Int

Create a new table; hash-cons as necessary

newtype SBVPgm

A program is a sequence of assignments

Constructors

SBVPgm

Fields

pgmAssignments :: Seq (SV, SBVExpr)

type Symbolic = SymbolicT IO

Symbolic is specialization of SymbolicT to the IO monad. Unless you are using transformers explicitly, this is the type you should prefer.

runSymbolic :: MonadIO m => SBVRunMode -> SymbolicT m a -> m (a, Result)

Generalization of runSymbolic

data State

The state of the symbolic interpreter

getPathCondition :: State -> SBool

Get the current path condition

extendPathCondition :: State -> (SBool -> SBool) -> State

Extend the path condition with the given test value.

inSMTMode :: State -> IO Bool

Are we running in proof mode?

data SBVRunMode

Different means of running a symbolic piece of code

Constructors

SMTMode QueryContext IStage Bool SMTConfig
In regular mode, with a stage. Bool is True if this is SAT.

CodeGen
Code generation mode.

Concrete (Maybe (Bool, [((Quantifier, NamedSymVar), Maybe CV)]))
Concrete simulation mode, with given environment if any. If Nothing: Random.

data Kind

Kind of symbolic value

Constructors

KBool

KBounded !Bool !Int

KUnbounded

KReal

KUninterpreted String (Either String [String])

KFloat

KDouble

KChar

KString

KList Kind

KSet Kind

KTuple [Kind]

KMaybe Kind

KEither Kind Kind

class Outputtable a where

A class representing what can be returned from a symbolic computation.

Methods

output :: MonadSymbolic m => a -> m a

Generalization of output

data Result

Result of running a symbolic computation

Constructors

Result

Fields

reskinds :: Set Kind

kinds used in the program

resTraces :: [(String, CV)]

quick-check counter-example information (if any)

resObservables :: [(String, CV -> Bool, SV)]

observable expressions (part of the model)

resUISegs :: [(String, [String])]

uninterpeted code segments

resInputs :: ([(Quantifier, NamedSymVar)], [NamedSymVar])

inputs (possibly existential) + tracker vars

resConsts :: [(SV, CV)]

constants

resTables :: [((Int, Kind, Kind), [SV])]

tables (automatically constructed) (tableno, index-type, result-type) elts

resArrays :: [(Int, ArrayInfo)]

arrays (user specified)

resUIConsts :: [(String, SBVType)]

uninterpreted constants

resAxioms :: [(String, [String])]

axioms

resAsgns :: SBVPgm

assignments

resConstraints :: Seq (Bool, [(String, String)], SV)

additional constraints (boolean)

resAssertions :: [(String, Maybe CallStack, SV)]

assertions

resOutputs :: [SV]

outputs

class SolverContext m where

Actions we can do in a context: Either at problem description time or while we are dynamically querying. Symbolic and Query are two instances of this class. Note that we use this mechanism internally and do not export it from SBV.

Minimal complete definition

constrain, softConstrain, namedConstraint, constrainWithAttribute, setOption, contextState

Methods

constrain :: SBool -> m ()

Add a constraint, any satisfying instance must satisfy this condition.

softConstrain :: SBool -> m ()

Add a soft constraint. The solver will try to satisfy this condition if possible, but won't if it cannot.

namedConstraint :: String -> SBool -> m ()

Add a named constraint. The name is used in unsat-core extraction.

constrainWithAttribute :: [(String, String)] -> SBool -> m ()

Add a constraint, with arbitrary attributes.

setInfo :: String -> [String] -> m ()

Set info. Example: setInfo ":status" ["unsat"].

setOption :: SMTOption -> m ()

Set an option.

setLogic :: Logic -> m ()

Set the logic.

setTimeOut :: Integer -> m ()

Set a solver time-out value, in milli-seconds. This function essentially translates to the SMTLib call (set-info :timeout val), and your backend solver may or may not support it! The amount given is in milliseconds. Also see the function timeOut for finer level control of time-outs, directly from SBV.

contextState :: m State

Get the state associated with this context

internalVariable :: State -> Kind -> IO SV

Create an internal variable, which acts as an input but isn't visible to the user. Such variables are existentially quantified in a SAT context, and universally quantified in a proof context.

internalConstraint :: State -> Bool -> [(String, String)] -> SVal -> IO ()

Require a boolean condition to be true in the state. Only used for internal purposes.

isCodeGenMode :: State -> IO Bool

Is this a CodeGen run? (i.e., generating code)

newtype SBVType

A simple type for SBV computations, used mainly for uninterpreted constants. We keep track of the signedness/size of the arguments. A non-function will have just one entry in the list.

Constructors

SBVType [Kind]

newUninterpreted :: State -> String -> SBVType -> Maybe [String] -> IO ()

Create a new uninterpreted symbol, possibly with user given code

addAxiom :: MonadSymbolic m => String -> [String] -> m ()

Generalization of addAxiom

data Quantifier

Quantifiers: forall or exists. Note that we allow arbitrary nestings.

Constructors

ALL

EX

needsExistentials :: [Quantifier] -> Bool

Are there any existential quantifiers?

data SMTLibPgm

Representation of an SMT-Lib program. In between pre and post goes the refuted models

Constructors

SMTLibPgm SMTLibVersion [String]

data SMTLibVersion

Representation of SMTLib Program versions. As of June 2015, we're dropping support for SMTLib1, and supporting SMTLib2 only. We keep this data-type around in case SMTLib3 comes along and we want to support 2 and 3 simultaneously.

Constructors

SMTLib2

smtLibVersionExtension :: SMTLibVersion -> String

The extension associated with the version

smtLibReservedNames :: [String]

Names reserved by SMTLib. This list is current as of Dec 6 2015; but of course there's no guarantee it'll stay that way.

data SolverCapabilities

Translation tricks needed for specific capabilities afforded by each solver

Constructors

SolverCapabilities

Fields

supportsQuantifiers :: Bool

Supports SMT-Lib2 style quantifiers?

supportsUninterpretedSorts :: Bool

Supports SMT-Lib2 style uninterpreted-sorts

supportsUnboundedInts :: Bool

Supports unbounded integers?

supportsReals :: Bool

Supports reals?

supportsApproxReals :: Bool

Supports printing of approximations of reals?

supportsIEEE754 :: Bool

Supports floating point numbers?

supportsSets :: Bool

Supports set operations?

supportsOptimization :: Bool

Supports optimization routines?

supportsPseudoBooleans :: Bool

Supports pseudo-boolean operations?

supportsCustomQueries :: Bool

Supports interactive queries per SMT-Lib?

supportsGlobalDecls :: Bool

Supports global declarations? (Needed for push-pop.)

supportsDataTypes :: Bool

Supports datatypes?

supportsFlattenedModels :: Maybe [String]

Supports flattened model output? (With given config lines.)

extractSymbolicSimulationState :: State -> IO Result

Grab the program from a running symbolic simulation state.

data SMTScript

A script, to be passed to the solver.

Constructors

SMTScript

Fields

scriptBody :: String

Initial feed

scriptModel :: [String]

Continuation script, to extract results

data Solver

Solvers that SBV is aware of

Constructors

Z3

Yices

Boolector

CVC4

MathSAT

ABC

data SMTSolver

An SMT solver

Constructors

SMTSolver

Fields

name :: Solver

The solver in use

executable :: String

The path to its executable

preprocess :: String -> String

Each line sent to the solver will be passed through this function (typically id)

options :: SMTConfig -> [String]

Options to provide to the solver

engine :: SMTEngine

The solver engine, responsible for interpreting solver output

capabilities :: SolverCapabilities

Various capabilities of the solver

data SMTResult

The result of an SMT solver call. Each constructor is tagged with the SMTConfig that created it so that further tools can inspect it and build layers of results, if needed. For ordinary uses of the library, this type should not be needed, instead use the accessor functions on it. (Custom Show instances and model extractors.)

Constructors

Unsatisfiable SMTConfig (Maybe [String])
Unsatisfiable. If unsat-cores are enabled, they will be returned in the second parameter.

Satisfiable SMTConfig SMTModel
Satisfiable with model

SatExtField SMTConfig SMTModel
Prover returned a model, but in an extension field containing Infinite/epsilon

Unknown SMTConfig SMTReasonUnknown
Prover returned unknown, with the given reason

ProofError SMTConfig [String] (Maybe SMTResult)
Prover errored out, with possibly a bogus result

data SMTModel

A model, as returned by a solver

Constructors

SMTModel

Fields

modelObjectives :: [(String, GeneralizedCV)]

Mapping of symbolic values to objective values.

modelBindings :: Maybe [((Quantifier, NamedSymVar), Maybe CV)]

Mapping of input variables as reported by the solver. Only collected if model validation is requested.

modelAssocs :: [(String, CV)]

Mapping of symbolic values to constants.

modelUIFuns :: [(String, (SBVType, ([([CV], CV)], CV)))]

Mapping of uninterpreted functions to association lists in the model. Note that an uninterpreted constant (function of arity 0) will be stored in the modelAssocs field.

data SMTConfig

Solver configuration. See also z3, yices, cvc4, boolector, mathSAT, etc. which are instantiations of this type for those solvers, with reasonable defaults. In particular, custom configuration can be created by varying those values. (Such as z3{verbose=True}.)

Most fields are self explanatory. The notion of precision for printing algebraic reals stems from the fact that such values does not necessarily have finite decimal representations, and hence we have to stop printing at some depth. It is important to emphasize that such values always have infinite precision internally. The issue is merely with how we print such an infinite precision value on the screen. The field printRealPrec controls the printing precision, by specifying the number of digits after the decimal point. The default value is 16, but it can be set to any positive integer.

When printing, SBV will add the suffix ... at the and of a real-value, if the given bound is not sufficient to represent the real-value exactly. Otherwise, the number will be written out in standard decimal notation. Note that SBV will always print the whole value if it is precise (i.e., if it fits in a finite number of digits), regardless of the precision limit. The limit only applies if the representation of the real value is not finite, i.e., if it is not rational.

The printBase field can be used to print numbers in base 2, 10, or 16. If base 2 or 16 is used, then floating-point values will be printed in their internal memory-layout format as well, which can come in handy for bit-precise analysis.

Constructors

SMTConfig

Fields

verbose :: Bool

Debug mode

timing :: Timing

Print timing information on how long different phases took (construction, solving, etc.)

printBase :: Int

Print integral literals in this base (2, 10, and 16 are supported.)

printRealPrec :: Int

Print algebraic real values with this precision. (SReal, default: 16)

satCmd :: String

Usually "(check-sat)". However, users might tweak it based on solver characteristics.

allSatMaxModelCount :: Maybe Int

In a allSat call, return at most this many models. If nothing, return all.

allSatPrintAlong :: Bool

In a allSat call, print models as they are found.

satTrackUFs :: Bool

In a sat call, should we try to extract values of uninterpreted functions?

isNonModelVar :: String -> Bool

When constructing a model, ignore variables whose name satisfy this predicate. (Default: (const False), i.e., don't ignore anything)

validateModel :: Bool

If set, SBV will attempt to validate the model it gets back from the solver.

optimizeValidateConstraints :: Bool

Validate optimization results. NB: Does NOT make sure the model is optimal, just checks they satisfy the constraints.

transcript :: Maybe FilePath

If Just, the entire interaction will be recorded as a playable file (for debugging purposes mostly)

smtLibVersion :: SMTLibVersion

What version of SMT-lib we use for the tool

solver :: SMTSolver

The actual SMT solver.

allowQuantifiedQueries :: Bool

Should we permit use of quantifiers in the query mode? (Default: False. See http://github.com/LeventErkok/sbv/issues/459 for why.)

roundingMode :: RoundingMode

Rounding mode to use for floating-point conversions

solverSetOptions :: [SMTOption]

Options to set as we start the solver

ignoreExitCode :: Bool

If true, we shall ignore the exit code upon exit. Otherwise we require ExitSuccess.

redirectVerbose :: Maybe FilePath

Redirect the verbose output to this file if given. If Nothing, stdout is implied.

data OptimizeStyle

Style of optimization. Note that in the pareto case the user is allowed to specify a max number of fronts to query the solver for, since there might potentially be an infinite number of them and there is no way to know exactly how many ahead of time. If Nothing is given, SBV will possibly loop forever if the number is really infinite.

Constructors

Lexicographic
Objectives are optimized in the order given, earlier objectives have higher priority.

Independent
Each objective is optimized independently.

Pareto (Maybe Int)
Objectives are optimized according to pareto front: That is, no objective can be made better without making some other worse.

data Penalty

Penalty for a soft-assertion. The default penalty is 1, with all soft-assertions belonging to the same objective goal. A positive weight and an optional group can be provided by using the Penalty constructor.

Constructors

DefaultPenalty
Default: Penalty of 1 and no group attached

Penalty Rational (Maybe String)
Penalty with a weight and an optional group

data Objective a

Objective of optimization. We can minimize, maximize, or give a soft assertion with a penalty for not satisfying it.

Constructors

Minimize String a
Minimize this metric

Maximize String a
Maximize this metric

AssertWithPenalty String a Penalty
A soft assertion, with an associated penalty

data QueryState

The state we keep track of as we interact with the solver

Constructors

QueryState

Fields

queryAsk :: Maybe Int -> String -> IO String

querySend :: Maybe Int -> String -> IO ()

queryRetrieveResponse :: Maybe Int -> IO String

queryConfig :: SMTConfig

queryTerminate :: IO ()

queryTimeOutValue :: Maybe Int

queryAssertionStackDepth :: Int

queryTblArrPreserveIndex :: Maybe (Int, Int)

newtype QueryT m a

A query is a user-guided mechanism to directly communicate and extract results from the solver. A generalization of Query.

Constructors

QueryT

Fields

runQueryT :: ReaderT State m a

newtype SMTProblem

Internal representation of a symbolic simulation result

Constructors

SMTProblem
SMTLib representation, given the config

Fields

smtLibPgm :: SMTConfig -> SMTLibPgm

Operations useful for instantiating SBV type classes

genLiteral :: Integral a => Kind -> a -> SBV b

Generate a finite constant bitvector

genFromCV :: Integral a => CV -> a

Convert a constant to an integral value

data CV

CV represents a concrete word of a fixed size: For signed words, the most significant digit is considered to be the sign.

Constructors

CV

Fields

_cvKind :: !Kind

cvVal :: !CVal

genMkSymVar :: MonadSymbolic m => Kind -> Maybe Quantifier -> Maybe String -> m (SBV a)

Generalization of genMkSymVar

genParse :: Integral a => Kind -> [CV] -> Maybe (a, [CV])

Parse a signed/sized value from a sequence of CVs

showModel :: SMTConfig -> SMTModel -> String

Show a model in human readable form. Ignore bindings to those variables that start with "__internal_sbv_" and also those marked as "nonModelVar" in the config; as these are only for internal purposes

data SMTModel

A model, as returned by a solver

Constructors

SMTModel

Fields

modelObjectives :: [(String, GeneralizedCV)]

Mapping of symbolic values to objective values.

modelBindings :: Maybe [((Quantifier, NamedSymVar), Maybe CV)]

Mapping of input variables as reported by the solver. Only collected if model validation is requested.

modelAssocs :: [(String, CV)]

Mapping of symbolic values to constants.

modelUIFuns :: [(String, (SBVType, ([([CV], CV)], CV)))]

Mapping of uninterpreted functions to association lists in the model. Note that an uninterpreted constant (function of arity 0) will be stored in the modelAssocs field.

liftQRem :: (Eq a, SymVal a) => SBV a -> SBV a -> (SBV a, SBV a)

Lift quotRem to symbolic words. Division by 0 is defined s.t. x/0 = 0; which holds even when x is 0 itself.

liftDMod :: (Ord a, SymVal a, Num a, SDivisible (SBV a)) => SBV a -> SBV a -> (SBV a, SBV a)

Lift divMod to symbolic words. Division by 0 is defined s.t. x/0 = 0; which holds even when x is 0 itself. Essentially, this is conversion from quotRem (truncate to 0) to divMod (truncate towards negative infinity)

registerKind :: State -> Kind -> IO ()

Register a new kind with the system, used for uninterpreted sorts. NB: Is it safe to have new kinds in query mode? It could be that the new kind might introduce a constraint that effects the logic. For instance, if we're seeing Double for the first time and using a BV logic, then things would fall apart. But this should be rare, and hopefully the success-response checking mechanism will catch the rare cases where this is an issue. In either case, the user can always arrange for the right logic by calling setLogic appropriately, so it seems safe to just allow for this.

Compilation to C, extras

compileToC' :: String -> SBVCodeGen a -> IO (a, CgConfig, CgPgmBundle)

Lower level version of compileToC, producing a CgPgmBundle

compileToCLib' :: String -> [(String, SBVCodeGen a)] -> IO ([a], CgConfig, CgPgmBundle)

Lower level version of compileToCLib, producing a CgPgmBundle

Code generation primitives

The codegen monad

newtype SBVCodeGen a

The code-generation monad. Allows for precise layout of input values reference parameters (for returning composite values in languages such as C), and return values.

Constructors

SBVCodeGen (StateT CgState Symbolic a)

cgSym :: Symbolic a -> SBVCodeGen a

Reach into symbolic monad from code-generation

Specifying inputs, SBV variants

cgInput :: SymVal a => String -> SBVCodeGen (SBV a)

Creates an atomic input in the generated code.

cgInputArr :: SymVal a => Int -> String -> SBVCodeGen [SBV a]

Creates an array input in the generated code.

cgOutput :: String -> SBV a -> SBVCodeGen ()

Creates an atomic output in the generated code.

cgOutputArr :: SymVal a => String -> [SBV a] -> SBVCodeGen ()

Creates an array output in the generated code.

cgReturn :: SBV a -> SBVCodeGen ()

Creates a returned (unnamed) value in the generated code.

cgReturnArr :: SymVal a => [SBV a] -> SBVCodeGen ()

Creates a returned (unnamed) array value in the generated code.

Specifying inputs, SVal variants

svCgInput :: Kind -> String -> SBVCodeGen SVal

Creates an atomic input in the generated code.

svCgInputArr :: Kind -> Int -> String -> SBVCodeGen [SVal]

Creates an array input in the generated code.

svCgOutput :: String -> SVal -> SBVCodeGen ()

Creates an atomic output in the generated code.

svCgOutputArr :: String -> [SVal] -> SBVCodeGen ()

Creates an array output in the generated code.

svCgReturn :: SVal -> SBVCodeGen ()

Creates a returned (unnamed) value in the generated code.

svCgReturnArr :: [SVal] -> SBVCodeGen ()

Creates a returned (unnamed) array value in the generated code.

Settings

cgPerformRTCs :: Bool -> SBVCodeGen ()

Sets RTC (run-time-checks) for index-out-of-bounds, shift-with-large value etc. on/off. Default: False.

cgSetDriverValues :: [Integer] -> SBVCodeGen ()

Sets driver program run time values, useful for generating programs with fixed drivers for testing. Default: None, i.e., use random values.

cgAddPrototype :: [String] -> SBVCodeGen ()

Adds the given lines to the header file generated, useful for generating programs with uninterpreted functions.

cgAddDecl :: [String] -> SBVCodeGen ()

Adds the given lines to the program file generated, useful for generating programs with uninterpreted functions.

cgAddLDFlags :: [String] -> SBVCodeGen ()

Adds the given words to the compiler options in the generated Makefile, useful for linking extra stuff in.

cgIgnoreSAssert :: Bool -> SBVCodeGen ()

Ignore assertions (those generated by sAssert calls) in the generated C code

cgOverwriteFiles :: Bool -> SBVCodeGen ()

If passed True, then we will not ask the user if we're overwriting files as we generate the C code. Otherwise, we'll prompt.

cgShowU8UsingHex :: Bool -> SBVCodeGen ()

If passed True, then we will show 'SWord 8' type in hex. Otherwise we'll show it in decimal. All signed types are shown decimal, and all unsigned larger types are shown hexadecimal otherwise.

cgIntegerSize :: Int -> SBVCodeGen ()

Sets number of bits to be used for representing the SInteger type in the generated C code. The argument must be one of 8, 16, 32, or 64. Note that this is essentially unsafe as the semantics of unbounded Haskell integers becomes reduced to the corresponding bit size, as typical in most C implementations.

cgSRealType :: CgSRealType -> SBVCodeGen ()

Sets the C type to be used for representing the SReal type in the generated C code. The setting can be one of C's "float", "double", or "long double", types, depending on the precision needed. Note that this is essentially unsafe as the semantics of infinite precision SReal values becomes reduced to the corresponding floating point type in C, and hence it is subject to rounding errors.

data CgSRealType

Possible mappings for the SReal type when translated to C. Used in conjunction with the function cgSRealType. Note that the particular characteristics of the mapped types depend on the platform and the compiler used for compiling the generated C program. See http://en.wikipedia.org/wiki/C_data_types for details.

Constructors

CgFloat
float

CgDouble
double

CgLongDouble
long double

Infrastructure

class CgTarget a where

Abstract over code generation for different languages

Methods

targetName :: a -> String

translate :: a -> CgConfig -> String -> CgState -> Result -> CgPgmBundle

data CgConfig

Options for code-generation.

Constructors

CgConfig

Fields

cgRTC :: Bool

If True, perform run-time-checks for index-out-of-bounds or shifting-by-large values etc.

cgInteger :: Maybe Int

Bit-size to use for representing SInteger (if any)

cgReal :: Maybe CgSRealType

Type to use for representing SReal (if any)

cgDriverVals :: [Integer]

Values to use for the driver program generated, useful for generating non-random drivers.

cgGenDriver :: Bool

If True, will generate a driver program

cgGenMakefile :: Bool

If True, will generate a makefile

cgIgnoreAsserts :: Bool

If True, will ignore sAssert calls

cgOverwriteGenerated :: Bool

If True, will overwrite the generated files without prompting.

cgShowU8InHex :: Bool

If True, then 8-bit unsigned values will be shown in hex as well, otherwise decimal. (Other types always shown in hex.)

data CgState

Code-generation state

Constructors

CgState

Fields

cgInputs :: [(String, CgVal)]

cgOutputs :: [(String, CgVal)]

cgReturns :: [CgVal]

cgPrototypes :: [String]

cgDecls :: [String]

cgLDFlags :: [String]

cgFinalConfig :: CgConfig

data CgPgmBundle

Representation of a collection of generated programs.

Constructors

CgPgmBundle (Maybe Int, Maybe CgSRealType) [(FilePath, (CgPgmKind, [Doc]))]

data CgPgmKind

Different kinds of "files" we can produce. Currently this is quite C specific.

Constructors

CgMakefile [String]

CgHeader [Doc]

CgSource

CgDriver

data CgVal

Abstraction of target language values

Constructors

CgAtomic SV

CgArray [SV]

defaultCgConfig :: CgConfig

Default options for code generation. The run-time checks are turned-off, and the driver values are completely random.

initCgState :: CgState

Initial configuration for code-generation

isCgDriver :: CgPgmKind -> Bool

Is this a driver program?

isCgMakefile :: CgPgmKind -> Bool

Is this a make file?

Generating collateral

cgGenerateDriver :: Bool -> SBVCodeGen ()

Should we generate a driver program? Default: True. When a library is generated, it will have a driver if any of the contituent functions has a driver. (See compileToCLib.)

cgGenerateMakefile :: Bool -> SBVCodeGen ()

Should we generate a Makefile? Default: True.

codeGen :: CgTarget l => l -> CgConfig -> String -> SBVCodeGen a -> IO (a, CgConfig, CgPgmBundle)

Generate code for a symbolic program, returning a Code-gen bundle, i.e., collection of makefiles, source code, headers, etc.

renderCgPgmBundle :: Maybe FilePath -> (CgConfig, CgPgmBundle) -> IO ()

Render a code-gen bundle to a directory or to stdout

Various math utilities around floats

fpMaxH :: RealFloat a => a -> a -> a

The SMT-Lib (in particular Z3) implementation for min/max for floats does not agree with Haskell's; and also it does not agree with what the hardware does. Sigh.. See: http://ghc.haskell.org/trac/ghc/ticket/10378 http://github.com/Z3Prover/z3/issues/68 So, we codify here what the Z3 (SMTLib) is implementing for fpMax. The discrepancy with Haskell is that the NaN propagation doesn't work in Haskell The discrepancy with x86 is that given +0/-0, x86 returns the second argument; SMTLib is non-deterministic

fpMinH :: RealFloat a => a -> a -> a

SMTLib compliant definition for fpMin. See the comments for fpMax.

fp2fp :: (RealFloat a, RealFloat b) => a -> b

Convert double to float and back. Essentially fromRational . toRational except careful on NaN, Infinities, and -0.

fpRemH :: RealFloat a => a -> a -> a

Compute the "floating-point" remainder function, the float/double value that remains from the division of x and y. There are strict rules around 0's, Infinities, and NaN's as coded below, See http://smt-lib.org/papers/BTRW14.pdf, towards the end of section 4.c.

fpRoundToIntegralH :: RealFloat a => a -> a

Convert a float to the nearest integral representable in that type

fpIsEqualObjectH :: RealFloat a => a -> a -> Bool

Check that two floats are the exact same values, i.e., +0/-0 does not compare equal, and NaN's compare equal to themselves.

fpCompareObjectH :: RealFloat a => a -> a -> Ordering

Ordering for floats, avoiding the +0-0NaN issues. Note that this is essentially used for indexing into a map, so we need to be total. Thus, the order we pick is: NaN -oo -0 +0 +oo The placement of NaN here is questionable, but immaterial.

fpIsNormalizedH :: RealFloat a => a -> Bool

Check if a number is "normal." Note that +0/-0 is not considered a normal-number and also this is not simply the negation of isDenormalized!

Pretty number printing

class PrettyNum a where

PrettyNum class captures printing of numbers in hex and binary formats; also supporting negative numbers.

Methods

hexS :: a -> String

Show a number in hexadecimal, starting with 0x and type.

binS :: a -> String

Show a number in binary, starting with 0b and type.

hexP :: a -> String

Show a number in hexadecimal, starting with 0x but no type.

binP :: a -> String

Show a number in binary, starting with 0b but no type.

hex :: a -> String

Show a number in hex, without prefix, or types.

bin :: a -> String

Show a number in bin, without prefix, or types.

readBin :: Num a => String -> a

A more convenient interface for reading binary numbers, also supports negative numbers

shex :: (Show a, Integral a) => Bool -> Bool -> (Bool, Int) -> a -> String

Show as a hexadecimal value. First bool controls whether type info is printed while the second boolean controls wether 0x prefix is printed. The tuple is the signedness and the bit-length of the input. The length of the string will not depend on the value, but rather the bit-length.

chex :: (Show a, Integral a) => Bool -> Bool -> (Bool, Int) -> a -> String

Show as hexadecimal, but for C programs. We have to be careful about printing min-bounds, since C does some funky casting, possibly losing the sign bit. In those cases, we use the defined constants in stdint.h. We also properly append the necessary suffixes as needed.

shexI :: Bool -> Bool -> Integer -> String

Show as a hexadecimal value, integer version. Almost the same as shex above except we don't have a bit-length so the length of the string will depend on the actual value.

sbin :: (Show a, Integral a) => Bool -> Bool -> (Bool, Int) -> a -> String

Similar to shex; except in binary.

sbinI :: Bool -> Bool -> Integer -> String

Similar to shexI; except in binary.

showCFloat :: Float -> String

A version of show for floats that generates correct C literals for nan/infinite. NB. Requires "math.h" to be included.

showCDouble :: Double -> String

A version of show for doubles that generates correct C literals for nan/infinite. NB. Requires "math.h" to be included.

showHFloat :: Float -> String

A version of show for floats that generates correct Haskell literals for nan/infinite

showHDouble :: Double -> String

A version of show for doubles that generates correct Haskell literals for nan/infinite

showSMTFloat :: RoundingMode -> Float -> String

A version of show for floats that generates correct SMTLib literals using the rounding mode

showSMTDouble :: RoundingMode -> Double -> String

A version of show for doubles that generates correct SMTLib literals using the rounding mode

smtRoundingMode :: RoundingMode -> String

Convert a rounding mode to the format SMT-Lib2 understands.

cvToSMTLib :: RoundingMode -> CV -> String

Convert a CV to an SMTLib2 compliant value

mkSkolemZero :: RoundingMode -> Kind -> String

Create a skolem 0 for the kind

Timing computations

data Timing

Specify how to save timing information, if at all.

Constructors

NoTiming

PrintTiming

SaveTiming (IORef NominalDiffTime)

showTDiff :: NominalDiffTime -> String

Show NominalDiffTime in human readable form. NominalDiffTime is essentially picoseconds (10^-12 seconds). We show it so that it's represented at the day:hour:minute:second.XXX granularity.

Coordinating with the solver

In rare cases it might be necessary to send an arbitrary string down to the solver. Needless to say, this should be avoided if at all possible. Users should prefer the provided API. If you do find yourself needing send and ask directly, please get in touch to see if SBV can support a typed API for your use case. Similarly, the function retrieveResponseFromSolver might occasionally be necessary to clean-up the communication buffer. We would like to hear if you do need these functions regularly so we can provide better support.

sendStringToSolver :: (MonadIO m, MonadQuery m) => String -> m ()

Send an arbitrary string to the solver in a query. Note that this is inherently dangerous as it can put the solver in an arbitrary state and confuse SBV. If you use this feature, you are on your own!

sendRequestToSolver :: (MonadIO m, MonadQuery m) => String -> m String

Send an arbitrary string to the solver in a query, and return a response. Note that this is inherently dangerous as it can put the solver in an arbitrary state and confuse SBV.

retrieveResponseFromSolver :: (MonadIO m, MonadQuery m) => String -> Maybe Int -> m [String]

Retrieve multiple responses from the solver, until it responds with a user given tag that we shall arrange for internally. The optional timeout is in milliseconds. If the time-out is exceeded, then we will raise an error. Note that this is inherently dangerous as it can put the solver in an arbitrary state and confuse SBV. If you use this feature, you are on your own!

Defining new metrics

addSValOptGoal :: MonadSymbolic m => Objective SVal -> m ()

Generalization of addSValOptGoal

sFloatAsComparableSWord32 :: SFloat -> SWord32

Convert a float to a comparable SWord32. The trick is to ignore the sign of -0, and if it's a negative value flip all the bits, and otherwise only flip the sign bit. This is known as the lexicographic ordering on floats and it works as long as you do not have a NaN.

sDoubleAsComparableSWord64 :: SDouble -> SWord64

Convert a double to a comparable SWord64. The trick is to ignore the sign of -0, and if it's a negative value flip all the bits, and otherwise only flip the sign bit. This is known as the lexicographic ordering on doubles and it works as long as you do not have a NaN.