Documentation.SBV.Examples.ProofTools.Sum

Description

Example inductive proof to show partial correctness of the traditional for-loop sum algorithm:

    s = 0
    i = 0
    while i <= n:
       s += i
       i++

We prove the loop invariant and establish partial correctness that s is the sum of all numbers up to and including n upon termination.

System state

data S a

System state. We simply have two components, parameterized over the type so we can put in both concrete and symbolic values.

Constructors

S

Fields

s :: a

i :: a

n :: a

sumCorrect :: IO (InductionResult (S Integer))

Encoding partial correctness of the sum algorithm. We have:

>>> sumCorrect
Q.E.D.