Documentation.SBV.Examples.Puzzles.MagicSquare

Description

Solves the magic-square puzzle. An NxN magic square is one where all entries are filled with numbers from 1 to NxN such that sums of all rows, columns and diagonals is the same.

Documentation

type Elem = SWord32

Use 32-bit words for elements.

type Row = [Elem]

A row is a list of elements

type Board = [Row]

The puzzle board is a list of rows

check :: Elem -> Elem -> [Elem] -> SBool

Checks that all elements in a list are within bounds

diag :: [[a]] -> [a]

Get the diagonal of a square matrix

isMagic :: Board -> SBool

Test if a given board is a magic square

chunk :: Int -> [a] -> [[a]]

Group a list of elements in the sublists of length i

magic :: Int -> IO ()

Given n, magic n prints all solutions to the nxn magic square problem