Documentation.SBV.Examples.Puzzles.Sudoku

Description

The Sudoku solver, quintessential SMT solver example!

Modeling Sudoku

type Row = [SWord8]

A row is a sequence of 8-bit words, too large indeed for representing 1-9, but does not harm

type Board = [Row]

A Sudoku board is a sequence of 9 rows

check :: [SWord8] -> SBool

Given a series of elements, make sure they are all different and they all are numbers between 1 and 9

valid :: Board -> SBool

Given a full Sudoku board, check that it is valid

type Puzzle = (Int, [SWord8] -> Board)

A puzzle is a pair: First is the number of missing elements, second is a function that given that many elements returns the final board.

Solving Sudoku puzzles

sudoku :: Puzzle -> IO ()

Solve a given puzzle and print the results

dispSolution :: Puzzle -> (Bool, [Word8]) -> IO ()

Helper function to display results nicely, not really needed, but helps presentation

solveAll :: Puzzle -> IO ()

Find all solutions to a puzzle

Example boards

puzzle0 :: Puzzle

Find an arbitrary good board

puzzle1 :: Puzzle

A random puzzle, found on the internet..

puzzle2 :: Puzzle

Another random puzzle, found on the internet..

puzzle3 :: Puzzle

Another random puzzle, found on the internet..

puzzle4 :: Puzzle

According to the web, this is the toughest sudoku puzzle ever.. It even has a name: Al Escargot: http://zonkedyak.blogspot.com/2006/11/worlds-hardest-sudoku-puzzle-al.html

puzzle5 :: Puzzle

This one has been called diabolical, apparently

puzzle6 :: Puzzle

The following is nefarious according to http://haskell.org/haskellwiki/Sudoku

allPuzzles :: IO ()

Solve them all, this takes a fraction of a second to run for each case