Documentation.SBV.Examples.Queries.AllSat

Description

When we would like to find all solutions to a problem, we can query the solver repeatedly, telling it to give us a new model each time. SBV already provides allSat that precisely does this. However, this example demonstrates how the query mode can be used to achieve the same, and can also incorporate extra conditions with easy as we walk through solutions.

Documentation

goodSum :: Symbolic [(Integer, Integer)]

Find all solutions to x + y .== 10 for positive x and y, but at each iteration we would like to ensure that the value of x we get is at least twice as large as the previous one. This is rather silly, but demonstrates how we can dynamically query the result and put in new constraints based on those.

demo :: IO ()

Run the query. We have:

>>> demo
Starting the all-sat engine!
Iteration: 1
Current solution is: (0,10)
Iteration: 2
Current solution is: (1,9)
Iteration: 3
Current solution is: (2,8)
Iteration: 4
Current solution is: (4,6)
Iteration: 5
Current solution is: (8,2)
Iteration: 6
No other solution!
[(0,10),(1,9),(2,8),(4,6),(8,2)]