Documentation.SBV.Examples.Queries.Enums

Description

Demonstrates the use of enumeration values during queries.

Documentation

data Day

Days of the week. We make it symbolic using the mkSymbolicEnumeration splice.

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

type SDay = SBV Day

The type synonym SDay is the symbolic variant of Day. (Similar to SInteger/Integer and others.)

findDays :: IO [Day]

A trivial query to find three consecutive days that's all before Thursday. The point here is that we can perform queries on such enumerated values and use getValue on them and return their values from queries just like any other value. We have:

>>> findDays
[Monday,Tuesday,Wednesday]