Documentation.SBV.Examples.Transformers.SymbolicEval

Description

A demonstration of the use of the SymbolicT and QueryT transformers in the setting of symbolic program evaluation.

In this example, we perform symbolic evaluation across three steps:

allocate free variables, so we can extract a model after evaluation

perform symbolic evaluation of a program and an associated property

querying the solver for whether it's possible to find a set of program inputs that falsify the property. if there is, we extract a model.

To simplify the example, our programs always have exactly two integer inputs named x and y.

Allocation of symbolic variables, so we can extract a model later.

newtype Alloc a

Monad for allocating free variables.

Constructors

Alloc

Fields

runAlloc :: SymbolicT (ExceptT String IO) a

data Env

Environment holding allocated variables.

Constructors

Env

Fields

envX :: SBV Integer

envY :: SBV Integer

result :: Maybe SVal

alloc :: String -> Alloc (SBV Integer)

Allocate an integer variable with the provided name.

allocEnv :: Alloc Env

Allocate an Env holding all input variables for the program.

Symbolic term evaluation

data Term :: * -> * where

The term language we use to express programs and properties.

Constructors

Var :: String -> Term r

Lit :: Integer -> Term Integer

Plus :: Term Integer -> Term Integer -> Term Integer

LessThan :: Term Integer -> Term Integer -> Term Bool

GreaterThan :: Term Integer -> Term Integer -> Term Bool

Equals :: Term Integer -> Term Integer -> Term Bool

Not :: Term Bool -> Term Bool

Or :: Term Bool -> Term Bool -> Term Bool

And :: Term Bool -> Term Bool -> Term Bool

Implies :: Term Bool -> Term Bool -> Term Bool

newtype Eval a

Monad for performing symbolic evaluation.

Constructors

Eval

Fields

unEval :: ReaderT Env (Except String) a

unsafeCastSBV :: SBV a -> SBV b

Unsafe cast for symbolic values. In production code, we would check types instead.

eval :: Term r -> Eval (SBV r)

Symbolic evaluation function for Term.

runEval :: Env -> Term a -> Except String (SBV a)

Runs symbolic evaluation, sending a Term to a symbolic value (or failing). Used for symbolic evaluation of programs and properties.

newtype Program a

A program that can reference two input variables, x and y.

Constructors

Program (Term a)

newtype Result

A symbolic value representing the result of running a program -- its output.

Constructors

Result SVal

mkResult :: SBV a -> Result

Makes a Result from a symbolic value.

runProgramEval :: Env -> Program a -> Except String Result

Performs symbolic evaluation of a Program.

Property evaluation

newtype Property

A property describes a quality of a Program. It is a Term yields a boolean value.

Constructors

Property (Term Bool)

runPropertyEval :: Result -> Env -> Property -> Except String (SBV Bool)

Performs symbolic evaluation of a 'Property.

Checking whether a program satisfies a property

data CheckResult

The result of checking the combination of a Program and a Property.

Constructors

Proved

Counterexample Integer Integer

generalize :: Monad m => Identity a -> m a

Sends an Identity computation to an arbitrary monadic computation.

newtype Q a

Monad for querying a solver.

Constructors

Q

Fields

runQ :: QueryT (ExceptT String IO) a

mkQuery :: Env -> Q CheckResult

Creates a computation that queries a solver and yields a CheckResult.

check :: Program a -> Property -> IO (Either String CheckResult)

Checks a Property of a Program (or fails).

Some examples

ex1 :: IO (Either String CheckResult)

Check that x+y+1 generates a counter-example for the property that the result is less than 10 when x+y is at least 9. We have:

>>> ex1
Right (Counterexample 0 9)

ex2 :: IO (Either String CheckResult)

Check that the program x+y correctly produces a result greater than 1 when both x and y are at least 1. We have:

>>> ex2
Right Proved

ex3 :: IO (Either String CheckResult)

Check that we catch the cases properly through the monad stack when there is a syntax error, like an undefined variable. We have:

>>> ex3
Left "unknown variable"