Documentation.SBV.Examples.WeakestPreconditions.Append

Description

Proof of correctness of an imperative list-append algorithm, using weakest preconditions. Illustrates the use of SBV's symbolic lists together with the WP algorithm.

Program state

data AppS a

The state of the length program, paramaterized over the element type a

Constructors

AppS

Fields

xs :: SList a

The first input list

ys :: SList a

The second input list

ts :: SList a

Temporary variable

zs :: SList a

Output

data AppC a

The concrete counterpart of AppS. Again, we can't simply use the duality between SBV a and a due to the difference between SList a and [a].

Constructors

AppC [a] [a] [a] [a]

type A = AppS Integer

Helper type synonym

The algorithm

algorithm :: Stmt A

The imperative append algorithm:

   zs = []
   ts = xs
   while not (null ts)
     zs = zs ++ [head ts]
     ts = tail ts
   ts = ys
   while not (null ts)
     zs = zs ++ [head ts]
     ts = tail ts

imperativeAppend :: Program A

A program is the algorithm, together with its pre- and post-conditions.

Correctness

correctness :: IO (ProofResult (AppC Integer))

We check that zs is xs ++ ys upon termination.

>>> correctness
Total correctness is established.
Q.E.D.