Documentation.SBV.Examples.WeakestPreconditions.IntDiv

Description

Proof of correctness of an imperative integer division algorithm, using weakest preconditions. The algorithm simply keeps subtracting the divisor until the desired quotient and the remainder is found.

Program state

data DivS a

The state for the division program, parameterized over a base type a.

Constructors

DivS

Fields

x :: a

The dividend

y :: a

The divisor

q :: a

The quotient

r :: a

The remainder

type D = DivS SInteger

Helper type synonym

The algorithm

algorithm :: Invariant D -> Maybe (Measure D) -> Stmt D

The imperative division algorithm, assuming non-negative x and strictly positive y:

   r = x                     -- set remainder to x
   q = 0                     -- set quotient  to 0
   while y <= r              -- while we can still subtract
     r = r - y                    -- reduce the remainder
     q = q + 1                    -- increase the quotient

Note that we need to explicitly annotate each loop with its invariant and the termination measure. For convenience, we take those two as parameters for simplicity.

pre :: D -> SBool

Precondition for our program: x must non-negative and y must be strictly positive. Note that there is an explicit call to abort in our program to protect against this case, so if we do not have this precondition, all programs will fail.

post :: D -> SBool

Postcondition for our program: Remainder must be non-negative and less than y, and it must hold that x = q*y + r:

noChange :: Stable D

Stability: x and y must remain unchanged.

imperativeDiv :: Invariant D -> Maybe (Measure D) -> Program D

A program is the algorithm, together with its pre- and post-conditions.

Correctness

invariant :: Invariant D

The invariant is simply that x = q * y + r holds at all times and r is strictly positive. We need the y > 0 part of the invariant to establish the measure decreases, which is guaranteed by our precondition.

measure :: Measure D

The measure. In each iteration r decreases, but always remains positive. Since y is strictly positive, r can serve as a measure for the loop.

correctness :: IO ()

Check that the program terminates and the post condition holds. We have:

>>> correctness
Total correctness is established.
Q.E.D.