Documentation.SBV.Examples.WeakestPreconditions.IntSqrt

Description

Proof of correctness of an imperative integer square-root algorithm, using weakest preconditions. The algorithm computes the floor of the square-root of a given non-negative integer by keeping a running some of all odd numbers starting from 1. Recall that 1+3+5+...+(2n+1) = (n+1)^2, thus we can stop the counting when we exceed the input number.

Program state

data SqrtS a

The state for the division program, parameterized over a base type a.

Constructors

SqrtS

Fields

x :: a

The input

sqrt :: a

The floor of the square root

i :: a

Successive squares, as the sum of j's

j :: a

Successive odds

type S = SqrtS SInteger

Helper type synonym

The algorithm

algorithm :: Invariant S -> Maybe (Measure S) -> Stmt S

The imperative square-root algorithm, assuming non-negative x

   sqrt = 0                  -- set sqrt to 0
   i    = 1                  -- set i to 1, sum of j's so far
   j    = 1                  -- set j to be the first odd number i
   while i <= x              -- while the sum hasn't exceeded x yet
     sqrt = sqrt + 1              -- increase the sqrt
     j    = j + 2                 -- next odd number
     i    = i + j                 -- running sum of j's

Note that we need to explicitly annotate each loop with its invariant and the termination measure. For convenience, we take those two as parameters for simplicity.

pre :: S -> SBool

Precondition for our program: x must be non-negative. Note that there is an explicit call to abort in our program to protect against this case, so if we do not have this precondition, all programs will fail.

post :: S -> SBool

Postcondition for our program: The sqrt squared must be less than or equal to x, and sqrt+1 squared must strictly exceed x.

noChange :: Stable S

Stability condition: Program must leave x unchanged.

imperativeSqrt :: Invariant S -> Maybe (Measure S) -> Program S

A program is the algorithm, together with its pre- and post-conditions.

Correctness

invariant :: Invariant S

The invariant is that at each iteration of the loop sqrt remains below or equal to the actual square-root, and i tracks the square of the next value. We also have that j is the sqrt'th odd value. Coming up with this invariant is not for the faint of heart, for details I would strongly recommend looking at Manna's seminal Mathematical Theory of Computation book (chapter 3). The j .> 0 part is needed to establish the termination.

measure :: Measure S

The measure. In each iteration i strictly increases, thus reducing the differential x - i

correctness :: IO ()

Check that the program terminates and the post condition holds. We have:

>>> correctness
Total correctness is established.
Q.E.D.