Documentation.SBV.Examples.WeakestPreconditions.Length

Description

Proof of correctness of an imperative list-length algorithm, using weakest preconditions. Illustrates the use of SBV's symbolic lists together with the WP algorithm.

Program state

data LenS a

The state of the length program, paramaterized over the element type a

Constructors

LenS

Fields

xs :: SList a

The input list

ys :: SList a

Copy of input

l :: SInteger

Running length

data LenC a

The concrete counterpart to LenS. Note that we can no longer use the duality between SBV a and a as in other examples and just use one datatype for both. This is because SList a and [a] are fundamentally different types. This can be a bit confusing at first, but the point is that it is the list that is symbolic in case of an SList a, not that we have a concrete list with symbolic elements in it. Subtle difference, but it is important to keep these two separate.

Constructors

LenC [a] [a] Integer

type S = LenS Integer

Helper type synonym

The algorithm

algorithm :: Invariant S -> Maybe (Measure S) -> Stmt S

The imperative length algorithm:

   ys = xs
   l  = 0
   while not (null ys)
     l  = l+1
     ys = tail ys

Note that we need to explicitly annotate each loop with its invariant and the termination measure. For convenience, we take those two as parameters, so we can experiment later.

pre :: S -> SBool

Precondition for our program. Nothing! It works for all lists.

post :: S -> SBool

Postcondition for our program: l must be the length of the input list.

noChange :: Stable S

Stability condition: Program must leave xs unchanged.

imperativeLength :: Invariant S -> Maybe (Measure S) -> Program S

A program is the algorithm, together with its pre- and post-conditions.

invariant :: Invariant S

The invariant simply relates the length of the input to the length of the current suffix and the length of the prefix traversed so far.

measure :: Measure S

The measure is obviously the length of ys, as we peel elements off of it through the loop.

Correctness

correctness :: IO ()

We check that l is the length of the input list xs upon termination. Note that even though this is an inductive proof, it is fairly easy to prove with our SMT based technology, which doesn't really handle induction at all! The usual inductive proof steps are baked into the invariant establishment phase of the WP proof. We have:

>>> correctness
Total correctness is established.
Q.E.D.